Create example_config.json
As the configured config.json file is in .gitignore for security reasons, it would be nice to know what information is needed in the config.json file in order to get this bot to run. This would make it so that anyone that is downloading this git repo would be able to replicate the necessary config.json file and have a working bot.